cpp/2019/02: rename files and Bazel targets
