+++ cpp/year2019/day02/BUILD.bazel
@@ -3,15 +3,15 @@ cc_library(
     hdrs = ["day02.h"],
     srcs = ["day02.cc"],
     deps = [
-        "@abseil//absl/strings",
+        "@abseil//absl/strings:strings",
         "//:adventofcode",
-        ":day02_internal",
+        ":internal",
     ],
 )
 
 cc_test(
-    name = "day02_test",
-    srcs = ["day02_test.cc"],
+    name = "test",
+    srcs = ["test.cc"],
     deps = [
         ":day02",
         "@googletest//:gtest_main",
@@ -20,24 +20,24 @@ cc_test(
 )
 
 cc_library(
-    name = "day02_internal",
-    hdrs = ["day02_internal.h"],
-    srcs = ["day02_internal.cc"],
+    name = "internal",
+    hdrs = ["internal.h"],
+    srcs = ["internal.cc"],
     visibility = [":__pkg__"],
 )
 
 cc_test(
-    name = "day02_internal_test",
-    srcs = ["day02_internal_test.cc"],
+    name = "internal_test",
+    srcs = ["internal_test.cc"],
     deps = [
-        ":day02_internal",
+        ":internal",
         "@googletest//:gtest_main",
     ],
 )
 
 cc_binary(
-    name = "day02_benchmark",
-    srcs = ["day02_benchmark.cc"],
+    name = "benchmark",
+    srcs = ["benchmark.cc"],
     deps = [
         ":day02",
         "@googlebenchmark//:benchmark_main",
